refactor: 컨트롤러 패키지 레이어 재구성
