+++ src/test/java/back/ecommerce/controller/CartControllerTest.java
@@ -34,6 +34,8 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import back.ecommerce.auth.annotaion.UserEmail;
+import back.ecommerce.controller.cart.CartController;
+import back.ecommerce.controller.common.GlobalExceptionHandler;
 import back.ecommerce.domain.product.Category;
 import back.ecommerce.dto.response.cart.CartProductDto;
 import back.ecommerce.dto.request.cart.AddCartRequest;
